Conversion corrections.
git-svn-id: https://svn.code.sf.net/p/spot-on/code/branches/0.13@3450 0a995953-517f-4c62-b348-104708d926b2

+++ Common/spot-on-crypt-ntru.cc
@@ -150,7 +150,7 @@ QByteArray spoton_crypt::publicKeyDecryptNTRU
       QByteArray privateKey;
       QByteArray publicKey;
 
-      privateKey.append(m_privateKey, m_privateKeyLength);
+      privateKey.append(m_privateKey, static_cast<int> (m_privateKeyLength));
       privateKey.remove(0, qstrlen("ntru-private-key-0000000000-"));
       memcpy(privateKey_array, privateKey.constData(), privateKey.length());
       ntru_import_priv(privateKey_array, &kp.priv);
